Share one ScriptedHost mock between the TUI agent unit tests
Problem: tui_agent_test.py's _ScriptedHost duplicated tui_utils_test.py's
_RecorderHost -- the same pydantic host stub (record commands, replay
scripted results, then succeed) defined twice under different names in
sibling test files.

Fix: moved the stub to a shared agents/mock_host_test.py as ScriptedHost
(following the existing mock_*_test.py convention for concrete interface
mocks) and updated both test files to import it.
